refactor: comment, notice, notification moves from middlewares to utils
